chore: move StartServer and StopServer from cmd to comp
