+++ pkg/cli/standalone/jmx.go
@@ -11,7 +11,6 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/DataDog/datadog-agent/cmd/agent/api"
 	internalAPI "github.com/DataDog/datadog-agent/comp/core/api"
 	logsAgent "github.com/DataDog/datadog-agent/comp/logs/agent"
 	"github.com/DataDog/datadog-agent/pkg/aggregator/sender"
@@ -58,7 +57,7 @@ func ExecJmxListWithRateMetricsJSON(selectedChecks []string, logLevel string, co
 // The common utils, including AutoConfig, must have already been initialized.
 func execJmxCommand(command string, selectedChecks []string, reporter jmxfetch.JMXReporter, output func(...interface{}), logLevel string, configs []integration.Config, senderManager sender.DiagnoseSenderManager, agentAPI internalAPI.Component) error {
 	// start the cmd HTTP server
-	if err := api.StartServer(nil, nil, nil, nil, nil, util.NewNoneOptional[logsAgent.Component](), senderManager, nil, nil); err != nil {
+	if err := agentAPI.StartServer(nil, nil, nil, nil, nil, util.NewNoneOptional[logsAgent.Component](), senderManager, nil, nil); err != nil {
 		return fmt.Errorf("Error while starting api server, exiting: %v", err)
 	}
